+++ src/main/java/kr/codesqaud/cafe/user/UserController.java
@@ -1,7 +1,13 @@
 package kr.codesqaud.cafe.user;
 
 import java.util.List;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
+import kr.codesqaud.cafe.login.LoginRequestDto;
+import kr.codesqaud.cafe.web.SessionConstant;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -13,6 +19,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 public class UserController {
 
     private final UserService userService;
+    private final Logger logger = LoggerFactory.getLogger(getClass());
 
     public UserController(UserService userService) {
         this.userService = userService;
@@ -45,5 +52,42 @@ public class UserController {
         User findUser = userService.findOne(userId).get();
         model.addAttribute("user", findUser);
         return "user/profile";
+    }
+
+    @GetMapping("/login")
+    public String loginPage() {
+        return "user/login";
+    }
+
+    @PostMapping("/login")
+    public String login(
+            @Valid LoginRequestDto loginRequestDto, BindingResult bindingResult, HttpServletRequest request) {
+        User loginUser = userService.login(loginRequestDto);
+
+        if (bindingResult.hasErrors()) {
+            logger.info("로그인 실패");
+            return "user/login_failed";
+        }
+
+        if (loginUser == null) {
+            logger.info("로그인 실패");
+            return "user/login_failed";
+        }
+        logger.info("로그인 성공");
+
+        HttpSession session = request.getSession();
+        session.setAttribute(SessionConstant.LOGIN_USER_ID, loginUser.getUserId());
+        session.setAttribute(SessionConstant.LOGIN_USER_NAME, loginUser.getName());
+
+        return "redirect:/";
+    }
+
+    @GetMapping("/logout") // TODO: Post로 리팩터링 필요
+    public String logout(HttpServletRequest request) {
+        HttpSession session = request.getSession(false);
+        if (session != null) {
+            session.invalidate();
+        }
+        return "redirect:/";
     }
 }
